address comments by skek
